more frontend improvements
